+++ src/Localisation/CLanguage.cpp
@@ -10,6 +10,8 @@ CLanguage* CLanguage::createWithJSON(matjson::Value value)
     pRet->displayNameNative = value["display_name_native"].asString().unwrapOr("Error");
     pRet->font = value["font"].asString().unwrapOr("notosans.fnt");
     pRet->useTrueTypeFallback = value["use-true-type-fallback"].asBool().unwrapOr(false);
+    pRet->splitDescByEveryChar = value["split-desc-by-every-char"].asBool().unwrapOr(false);
+    pRet->dontTouchDescriptions = value["dont-split-descriptions"].asBool().unwrapOr(false);
 
     std::string usage = value["alt-font-usage"].asString().unwrapOr("Auto");
 
@@ -126,4 +128,14 @@ const int& CLanguage::getFontUsageMode()
 const bool& CLanguage::getTrueTypeFallback()
 {
     return useTrueTypeFallback;
+}
+
+const bool& CLanguage::getSplitEachChar()
+{
+    return splitDescByEveryChar;
+}
+
+const bool& CLanguage::getDontTouchDescriptions()
+{
+    return dontTouchDescriptions;
 }